images: keep the image-generation claim out of capability identity; fit the line budget
decode_images moves to openai_protocol/images_requests.py; existing traces and
frozen catalogs keep their digests (the claim is admitted fail-closed on its
own surface, like supports_completions).
