+++ exp/common/models/model_test.py
@@ -105,6 +105,7 @@ def test_model_request_keeps_tool_contract_and_capabilities_deterministic() -> N
     assert ModelCapabilities(supports_tools=True).model_dump(mode="json") == {
         "supports_tools": True,
         "supports_embeddings": None,
+        "supports_image_generation": None,
         "supports_structured_output": False,
         "supports_completions": None,
         "supports_temperature": True,
